Make types be highlighted consistently
Without this update types are sometimes highlighted and sometimes not.
So when you declare a new type (`type T …` the type name will be
highlighted, but if you then refer to the type later in your code, it
will not be highlighted.

This feels a little weird and non-consistent and other editors (like
Sublime) have chosen to not highlight the type name while declaring so
that the formatting of the type name is consistent across your code.