Updating/fixing the rune matching
The class `constant.rune` is generally available in theme as the name
`rune` is more of less Go language specific. So renamed the class to
`constant.other.rune` which then is matched against `constant.other`
which does seem to be general available.

Also added some more testing and matching to invalidate any other uses
of single quoted strings following the language specs.