Update keyword.operator.comparison.go regex
The regex to match keyword.operator.comparison.go
(==|!=|<=|>=|<[^<]|>[^>])
contains a mistake introduced to avoid matching the bitshift operators. [^<] and [^>] will match anything following the first less/greater than symbol (this highlights <3 and >3 and is even more obvious with <3.0 or >2.0 
The correct regex is
(==|!=|<=|>=|<(?!<)|>(?!>))
as it uses negative lookaheads to avoid highlighting characters immediately following the less/greater than symbol.