Here's a update based on Autumn's suggestions
There's still a lot of changes needed eventually to this to reroute things because it's going to be confusing going further into it and still seeing things linked to the old Atom branding for now when things might be better off in this repository. I at least made the wording changes Autumn suggested.