Update Cirrus Badge link
Update link to specifically point to the master branch. The badge itself is fine, as master is the base branch, and is where shields.io already points to

However, the link is a slightly different story as if you don't specify a branch, it shows all runs for the entire repo, which could be detrimental to new users, and possibly lead to non-standard or test installations, instead of the "release-ready" master build.