Fix File->Open not responding
When an user click fon File->Open, the event application:open is sent.
It's handled in 2 places: atom-application and
register-default-commands. Because getDefaultPath() requires
AtomEnvironment which isn't available in the main process with Electron 12, this handler
will fail. The fix is to remove that handler so the default handler will
be used.