Log about benchmark removal before main process handoff
If a new main process is launched with --benchmark mode, we have to
log quite early in the startup process, because it pretty quickly
notices the existing main process (by detecting its socket server),
and hands control of whether/how to launch any new windows to the
original, existing main process that was already running.

The new main process terminates immediately after passing the
message to the original main process about what it was trying to do.

(Particularly: it sends the first main process an encoded copy
of the options that were passed to the newer main process at launch.)

So we have to log something in the new main process quite early,
before the handoff occurs and the new main process terminates.

Otherwise, the newer terminal view won't show any info about
benchmark mode being removed, and it might not be obvious
why no new window was launched.