meta: Don't exclude 'loophole' or 'pegjs' packages
We need these when typing a "tab" character in a document or file.

(Without 'loophole' and 'pegjs' in the app bundle,
'snippets' tries and fails to require() those packages,
and it errors out every time you type a tab character.)

Explanation:

Our PegJS files _aren't_ pre-compiled at the moment.
We removed the old build scripts that used to transpile these files
during the build, and without replacing that functionality yet.

So...

When a certain PegJS --> JS file isn't available in the 'snippets'
package (because it hasn't been transpiled to plain JS yet yet),
'snippets' require()s some modules that let it live-transpile
the needed code, or at least that's what I think it's doing
in that scenario.

If the modules it conditionally require()s in that scenario aren't
in the Pulsar bundle, due to being explicitly excluded, then we get a
require() error, since they legitimately aren't in the bundle.