Revert "ci: Set internal version to end in '0-beta', no date"
This reverts the immediately previous commit
44f11c1d1f01b85860f8752b98242199fde94ddb.

Reverting, but keeping it in the commit history as a template
for how we can get cirrus to make binaries for a specific
release suffix. For example `-beta` or `-alpha`, and so on.

See the previous commit (the one this reverts) for an example of how.