src: Same symlink names, no "release channels"
(Also fix the link "apm" to instead be "ppm", per our rebranding.)

Uses the same symlink names, regardless of Stable/Beta/Nightly/Dev
"release channels".

These are detected at run-time by inspecting the version string.

If the version string ends in "-beta" it is considered a "Beta"
channel release. For "-dev", it is a "Dev" release. For "-nightly*",
it is considered a "Nightly" release. For a bare number, it is
considered a "Stable" release.

We don't actually have build or release infrastructure to have the app
actually behave differently as was originally intended in Atom
based on these "release channels".

We just have "Pulsar"/"pulsar" (case) and "ppm" at the moment.
No "pulsar-dev", "ppm-beta", "Pulsar Nightly", etc. etc. etc.

Although the app title may still display as e.g. "Pulsar Beta",
so we need to be careful to correct the app's internal behavior
to not malfunction for these cases where the "release channel"
(a thing we don't really do at the moment) is detected as non-stable.

FWIW, most of our builds off of `master` branch are simply detected
as "stable" release channel.