script: Check ppm symlink, not its target
Use `test -L` rather than `test -f` for `post-install.sh`.

(Note: `[ ]` is an alternate invocation of the `test` shell builtin.)

`[ -f some_path ]`, for a symlink "dereferences" the symlink, meaning
it checks if its target (or the ultimate target, in the case of
a chain of symlinks) exists and is a real file.

By coincidence, as this runs after we extract/install ppm from the deb
or rpm, the symlink target exists and is a real file. However, if we
changed the install path for our package manager, the target might not
exist, and this step might fail.

`[ -L some_path ]` instead checks for if the path exists and is a
symlink. This is more correct -- it's more directly what we're trying
to check.

So we will use `-L` rather than `-f`.