deps: Update github to v0.36.14-pretranspiled-take-2
Stops implicitly instructing the editor to transpile github, since
this tagged version of the package is pretranspiled.

Transpiling again would be redundant, presumably would waste CPU
cycles, and can show errors if devDependencies are missing. (And they
*would* be missing, due to the way this package is currently
specified in package.json, as a *tarball*, not a git ref URL where the
devDependencies *would* be included.)