autocomplete-css: Put implicit values last
(This commit should just be re-arranging the values,
so as to put the implicit ones last.)

These implicit values are the same for all properties.

I find it more useful to view the property-specific values first,
since users may not actually know what property-specific values exist
for a given CSS property, and these implicit values are *always*
allowed... (I think they are, anyway).

This shows those property-specific values first, better advertising
what options are actually available to use. (Besides that, as far as I
know, the implicit "inherit", "initial", "unset" values are not as
commonly used as the property-specific values anyway.)