autocomplete-css: Handle recursion for appending implicit values
Avoid appending the implicit values multiple times / in multiple
spots in the values list per property.

(These duplicate implicit values would be deduped, but the middle ones
would be the ones that would be kept, whereas we want to only keep
the bottom ones.)

---

Details:

The function that adds the implicit values to each property is
recursed sometimes, based on the structure of the data we are
ingesting. For the deeper/innner calls of this function, avoid
appending the implicit values, since we will be doing that at the end
of the outer/upper calls.

Makes sure the implicit values are not added in the middle *and* at
the bottom, in which scenario they would eventually be deduped
(removing the bottom copy, I suppose), but leaving one copy of the
implicit values smack-dab in the middle of the values list for the
given property.

We want only one copy of these values, sure, so deduping is
appreciated, but they should be *at the bottom* of the list of values.
Not the middle. So we gotta make sure the only copy fed through the
deduper function is the one at the bottom.