Changed the matching of scopes
A quick explanation on this - the old code was matching if the full
scope was ok. This is fine when we want to test a single grammar, but
for example, TextMate grammars always added `.ruby` on the last part of
the grammar. That was quite bad for TreeSitter, that basically didn't do
the same thing. So now it matches a fine crafted Regexp that basically
checks if the full scope is match from the beginning, OR if part of the
scope (up to just before the `.`) matches. So, for example, for
`constant.other.ruby` it'll match `constant`, `constant.other`, and
`constant.other.ruby` but it'll NOT MATCH `constant.oth`.