Back to using languageMode.ready
Just to document this: I tried a HUGE amount of ideas for tests to run
reliably. They need to wait for the editor when it's ready to be
tokenized, but our callbacks are not that reliable - meaning, for the
first tokenization on tree-sitter, they are called _after_ we register
the callback, but for the second time, they are called _before_ we
register them. There's no reliable way to _actually listen_ to changes,
and none of these even were _called_ by TextMate grammars.

So, this `.ready` will stay here so we know when TreeSitter is ready to
tokenize.