Fix `detectCoveredScope`
In EJS I ran into a situation that clearly was a false positive for the
`detectCoveredScope` test.

I am almost certain now that this should not be left up to a heuristic;
there should just be a `coverParentScope` option in
`atom.grammars.addInjectionPoint` so that it can be handled on a
case-by-case basis.

Regex injections will want to cover parent scopes; the `language-todo` and
`language-hyperlink` injections likely will not. And if an injection wants
to cover the parent, it should cover the parent _entirely_, not just when
their boundaries happen to line up.