Move folding logic to a `FoldResolver`…
…and add “divided” folds that specify only one end of a possible fold.

Divided folds aren't fun, but they make the existing tests pass, because some kinds of fold could not be expressed without them. They force us to scan almost the entire buffer just to be able to find the end of a particular fold, but the performance has not been a problem yet, and probably will not be. On the other hand, I would be _very_ reluctant to allow a similar approach for syntax highlighting, because I fear it could become a bottleneck and a footgun.