Fix getRangeForSyntaxNode predicate usage
The `bracket-matcher` package was assuming that `getSyntaxNodeForPosition`'s second argument — the test predicate — would stop running as soon as we found our match, and was using it to assign `startTag` and `endTag` variables. This is a fair assumption to make. So we shouldn't run the predicate until we're prepared to return immediately as soon as we find a match. That means putting _all_ candidates into a list, sorting that list from smallest to largest, and only then running the predicate.