Streamline behavior around when injections include their root scope
An injection layer will include its root scope over its entire extent — i.e., whatever node is the root of its injection point — by default.

You can change the injected language's scope name with the new `languageScope` option on `addInjectionPoint`, but more useful is the ability to set it to `null` and bypass this behavior.

That's the right thing to do for the builtin injections `text.todo` and `text.hyperlink`, or else their scopes would always be present.

(This is a pain in the ass and I'm not 100% sure we've found the right solution. Pretend there's an injection for an embedded language like ERB; it'd inject into a bunch of disparate ranges, but still apply a scope over the entire extent, even the ranges where it's not active. The alternative — defining start/end scope boundaries for each of the disjoint ranges — is pretty chaotic in its own right, since those ranges tend to ignore things like whitespace. I would want such a language to set `languageScope: null`, then set its root scope via capture at whatever level of granularity makes sense for that language.)