Laying the ground for indentation overhaul
I spent a lot of time pursuing various things we could try to cut down on the amount of tree reparsing we'd need.

I haven't committed any of it, but those experiments did inform some changes that are present here. When wanting to perform an operation that would require a tree re-parse, we can now theoretically choose between (a) re-parsing and staying synchronous or (b) waiting until the next scheduled tree parse and going async.

The implications for how indentation works are too big to plow ahead with this approach, but it would be worth experimenting with. The vast majority of all our indentation decisions are prompted by a single keystroke, and in those cases there's no particular reason why we can't wait until a few milliseconds later when we're going to be re-parsing anyway.