Add `suggestedIndentForBufferRows`…
…for suggesting the indent of a number of buffer rows at once.

The way `TextEditor` wants to auto-indent is by asking each row what its
indent level should be, then indenting each one in turn. This is because it
knows the language mode will probably determine row X's correct indentation by
using row X-1's indentation as a baseline, so there's no point in even calling
`suggestedIndentForBufferRow` for line X until X-1 is correctly indented.

This isn't ideal for us because each editor change will prompt a tree re-parse.
Instead, `suggestedIndentForBufferRows` will return the correct indentation
level for each row in a range without making any buffer changes.

It can do this because it tells `suggestedIndentForBufferRow` what the previous
row's indentation level is – hypothetically — so that it doesn't try to look it
up. Only the first row in the range compares its indentation to the _actual_
indentation of its preceding row; the rest of the rows figure out their
indentations relative to the hypothetical value we computed in the previous trip
through the loop.

This is paired with special logic in `TextEditor` to use this method instead of
its typical code path when asked to auto-indent multiple rows at once. Thus
we'll reduce the number of unscheduled tree parses from N (where N is the total
number of rows wanting to be auto-indented) to either zero or one.

Amazingly, I think this is the first time I've had to touch `text-editor.js` for
any reason.