remove this todo
imagine a scenario, a cache just got loaded, it contains keys for strings that no longer exist, and the app still requests those strings, and then the cache would return it? this would be a bug that's hidden, since we would want to know that the string is missing. Currently we look for the real thing first, so this bug doesn't exist