Add scope tests and specs
Added and renamed a few specs, most notably the new `onlyIfConfig`.

Wrote specs for all the positive versions of tests, omitting specs for the
negations; therefore I've aimed to make the negations as simple as possible so
that they are implicitly proven to work as long as their counterparts work.