Update `suggestedIndentForBufferRows`…
…to support `@match` captures.

Instead of just telling `suggestedIndentForBufferRow` (singular) the indent
level of the previous row, we should tell it the (hypothetical) indent levels
of all previous rows that are part of the indentation batch. That way it can
use that table when resolving a `@match` capture.

If the row to be matched is within the batch indent job, it can read the
indentation level from the table. If it's outside of the batch area, it can
safely read the row's indentation level directly.

I suspected it was a good idea to require `@match` captures to refer to rows
_above_ them, rather than any arbitrary row in the buffer, but I don't think I
had a specific rationale at the time. Now I'm glad I did it, whatever the
reason.