Fix insane amounts of tree re-parsing
This is partly something that I copied directly from `TreeSitterLanguageMode`,
but I don't think it's ever quite worked right.

First, the part that's entirely my fault: in trying to keep track of when the
tree is dirty and when it isn't — to at least bring our attention to how often
we're doing off-schedule parses — I marked each `LanguageLayer` with
`this.treeIsDirty = true` the first time `handleTextChange` ran after a
transaction, and would mark them with `this.treeIsDirty = false` after the tree
parse at the end of a transaction.

But not every layer runs `update` after every transaction — only the ones whose
markers were encroached on by the buffer changes! It would _appear_ that as long
as those distant layers are kept up-to-date on changes via `Tree#edit`, they
don't really need to be re-parsed immediately, and can wait until a change that
would actually affect their contents.

At any rate, since some layers didn't get calls to `update`, they also didn't
have their `treeIsDirty` flag reset, and thought that they had dirty trees even
at rest. Combined with my recent realization that off-schedule parses can't be
retained on injection layers, this resulted in a _whole lot of_ “re-parsing”
warnings, even at rest. Even simply when _scrolling_ the editor.

Now, at the end of a transaction, we'll loop through the current injection
layers and set `treeIsDirty` to `false` to be sure.

The other part that doesn't seem right is present in `TreeSitterLanguageMode`:
every single language layer gets `handleTextChange` called with every single
buffer edit (as is necessary), but every single layer keeps track of the
accumulated changes via the `editedRange` property — a range that can only ever
grow over time _until_ the layer's tree is re-parsed. And since we've
demonstrated that some layers may not actually get re-parsed until much later,
we can end up with a situation where (e.g.) a two-line injection on line 500
will eventually invalidate 1000 lines of highlighting _if_ the user makes two
edits: one on line 1 and one on line 1000. Even if they're minutes apart in
separate transactions.

Thus I'm nearly certain that this is wrong as well. If the edits of a
transaction do not result in a given layer being re-parsed, then the fact that a
certain range of the buffer was edited in that transaction ceases to be
relevant, and certainly should not be grown to include the ranges of changes in
_later transactions_. Hence: at the end of a transaction, all language layers
should also have their `editedRange` property set to `null` (if it wasn't
already).

I haven't investigated yet, but I suspect that it _used_ to be the case that
every layer was re-parsed on every transaction. That would explain why this
may have once made sense.

In other news, my idea for the foldable cache to contain ranges rather than
booleans… was silly. The cache is smart enough to shift lines around along with
buffer changes… but for cached ranges to be useful, we'd have to translate the
ranges along with each buffer change, and that's just too much work, especially
when fold captures are so cheap.