Fix highlighting for `coverShallowerScopes`
All the logic that we had employed to decide whether an injection ought to
prevent a shallower layer from scoping a particular region… was completely
ignored when determining which scopes should already be open as part of
`HighlightIterator#seek`. Hence a scope boundary might be suppressed if it's in
the middle of a highlighting job's range, but still might be assumed to be open
for a highlighting job that starts a few lines later.

To fix this, we've got to keep track of when those already-open scopes would've
opened, and if we find that one injection layer has enabled
`coverShallowerScopes`, we must re-assess those open scopes and remove the ones
that _would've_ been suppressed if we were highlighting the area where the scope
would've opened.

With me so far?

There's another thing that needed fixing: an injection layer can stop its parent
from applying a scope to an arbitrary token, but it should not be able to stop
its parent from applying its _base_ scope, like `source.js`. That's also been
fixed.

If I had to do it all over again, I would've dropped the ability of one layer to
“cover” the scopes of another, but it was already half-implemented in
`TreeSitterLanguageMode` and I felt compelled to implement it _properly_,
for whatever definition of “properly” I had in my head at the time. Now it's an
opt-in behavior, and anyone who opts into it deserves what they get.

All the existing specs still pass, and I added a new one for this new behavior.
Luckily, the current implementation of `scopeDescriptorForPosition` is the
perfect way to test this, because it ends up calling `HighlightIterator#seek` at
the precise position we tell it to.

And a new option for injection points has been added:
`includeAdjacentWhitespace`. During my windmill-tilting efforts to take an
otherwise great `tree-sitter-markdown` parser and try to inject a YAML
front-matter section — despite its lack of understanding of that syntax — I
discovered that a certain section of YAML wasn't parsing right because the
`tree-sitter-markdown` parser was ignoring a bit of leading whitespace that the
`tree-sitter-yaml` parser needed to see, because it was syntactically
meaningful. That made me realize how to handle this in the injection point API.

I wouldn't dare try to land that YAML/Markdown hack into the Pulsar core, but
the `includeAdjacentWhitespace` option will be useful for other injections, most
notably PHP. When a `content` callback returns an array of nodes, often that
gets converted into a series of disjoint ranges that _aren't_ actually disjoint;
they're just consecutive syntax nodes separated by whitespace. So the new option
takes this into account; when it's enabled, if two would-be injection ranges are
separated only by whitespace, then those two ranges are consolidated into a
single range _along with_ that whitespace.