Dispose of old tree-sitter trees when they're no longer needed.
I'd tried this once before and had it cause occasional errors, but the
tree-sitter playground deletes old trees itself, so I'm assuming I made a
mistake.

The lesson of `vscode-oniguruma` is that we need to know how a WASM dependency
manages memory, and though `web-tree-sitter` is very good about this internally,
the tree-sitter playground's example makes it clear that we should be manually
GC-ing old trees.

To be sure that we're only GC-ing trees that we don't need anymore, I've made it
so that tree deletion only happens after transactions, even for trees that
result from off-schedule parses.