Attempt to keep an injection layer's ranges fresh…
…even during mid-transaction buffer changes.

If we're going to use off-schedule tree parses for any purpose, they should at
least be accurate enough to react to the user's typing, instead of assuming that
the last known injection ranges are still valid.

So we create a marker for each range in order to allow those ranges to react to
buffer changes by growing, shrinking, or shifting, as the situation warrants.

This still doesn't guarantee we're parsing the correct ranges, but it's
certainly more accurate than assuming the injection ranges won't change at all
in between transactions.

I'd include a spec for this behavior if I had _any idea_ how to test this
properly.