Better fix: ensure `lastSyntaxTree` is up-to-date
The comparison described in the previous commit works just fine — it just needs
us to ensure `lastSyntaxTree` continues to receive edits to keep it up-to-date
with buffer changes. We were failing to do that before because the off-schedule
tree got those edits _instead of_, rather than _in addition to_, the tree from
the previous transaction.