Add experimental support for async indentation
The goal is to support as much as possible of what can be done synchronously
in TextMate grammars (using a perplexing regex-based system) by going async when
necessary in modern Tree-sitter grammars (using an `indents.scm` file that is
far easier to understand).

The one true concession we have to make here is to accept that certain
auto-indent behaviors can't be replicated atomically if they happen inside of a
transaction alongside other changes. In those cases, we must wait until the end
of the transaction, then auto-indent the entire range affected by the
transaction instead.

If you're interested in knowing more about this… well, I'd be surprised. But the
entire rationale is documented here:

https://gist.github.com/savetheclocktower/4fd4e3edce76ee5820990db0f012062e