Support “editor:paste-without-formatting“…
…via the same machinery as our post-insertion auto-indent.

The `preserveTrailingLineIndentation` option is confusingly named. Its purpose
is to say this: “Don't auto-indent every line in the pasted text. Insert the
first line at the proper indent level for the context, but otherwise preserve
the relative indentation of each line.” This is what the
`editor:paste-without-formatting` command does.

We already have a workflow where `TextEditor` asks a language mode for the
correct indent levels for a range of lines, so we can add a mode to
`WASMTreeSitterLanguageMode#suggestedIndentForBufferRows` that implements the
logic of `preserveTrailingLineIndentation` and reports indent levels
accordingly.