Avoid unnecessary injection repopulation…
…by ignoring any grammars that can't affect our injections.

This applies during startup; if a grammar that could inject into an open file
is added _after_ the open file is parsed and highlighted, we must repopulate
injections just in case the new grammar is applicable. But we should not react
to non-WASMTreeSitter grammars, nor grammars that do not define an
`injectionRegex`.