Detect highlighting tasks when the tree is dirty…
…and schedule them to be redone when the tree is clean.

Async parsing means we have to account for this. We tell the display layer when
ranges need to be re-highlighted, but we can't be sure when those tasks will be
scheduled, or if the tree will be clean when we do the required `highlights.scm`
captures.

We have several options:

1. When we run the query and notice that the tree is dirty, do an off-schedule
   synchronous parse.

2. When `HighlightIterator#seek` notices the tree is dirty, cancel the highlight
   job and reschedule it for whenever the tree is clean.

3. When `HighlightIterator#seek` notices the tree is dirty, proceed with the
   highlight job but _also_ schedule the same job to run again when the tree is
   clean.

#1 is tempting, but ultimately irresponsible if we don't know how long that
synchronous parse will take.

#2 is not technically possible, because the highlighting job will proceed no
matter what we do, so if we return early from `seek`, the display layer will
interpret that to mean the entire buffer range is plain text.

#3 is what I've implemented here. Captures against a dirty tree are likely to
be at least partially inaccurate, but since we have to go through with the job
anyway, we might as well allow it to proceed, thereby minimizing the flash of
unhighlighted text.

I can't think of a spec that will catch this situation reliably, but I'll give
it a shot later on.